refactor(DocumentController): combine Markdown and Mapping controller
