+++ server/utils/common.ts
@@ -2,21 +2,26 @@ import path from 'path';
 import fs from 'fs-extra';
 import { error } from './log';
 import { format } from 'date-fns';
-import { ExHentaiInfoItem } from '../controller/ExhentaiController';
 
 export const joinWithRootPath = (url: string) => path.join(process.cwd(), url);
 
-export const writeIntoJsonFile = (url: string, content: any) =>
-  fs
-    .outputJSON(path.join(process.cwd(), url + '.json'), content)
-    .catch((err: Error) => {
-      error(`error in utils/common.writeInfoJsonFile => ${url}: ${err}`);
-      throw err;
-    });
+export const writeIntoJsonFile = (
+  url: string,
+  content: any,
+  spaces?: boolean,
+) => {
+  if (!url.endsWith('.json')) {
+    url += '.json';
+  }
+  fs.outputJSON(path.join(process.cwd(), url), content, {
+    spaces: spaces ? 2 : 0,
+  }).catch((err: Error) => {
+    error(`error in utils/common.writeInfoJsonFile => ${url}: ${err}`);
+    throw err;
+  });
+};
 
-export const readJsonFile = (
-  url?: string | undefined | null,
-): ExHentaiInfoItem[] => {
+export const readJsonFile = (url?: string | undefined | null): any => {
   if (!url) {
     error('url is undefined.');
     throw Error;
