+++ src/controller/MainPageDataController.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import MainPage from '../pages/MainPage';
-import { MappingProps } from '../../server/controller/MarkdownController';
+import { MappingProps } from '../../server/controller/DocumentController';
 import { FormProps } from '../pages/EditForm';
 import { message } from 'antd';
 
@@ -53,7 +53,7 @@ export default class MainPageDataController extends Component<
   };
 
   handleDelete = async ({ id, category }: { id: string; category: string }) => {
-    const response = await fetch('/del/mapping', {
+    const response = await fetch('/document/delete', {
       method: 'DELETE',
       body: JSON.stringify({ id, category }),
       mode: 'cors',
@@ -82,14 +82,14 @@ export default class MainPageDataController extends Component<
     let id;
     if (dataItem) {
       id = dataItem.id;
-      await fetch('save/update', {
+      await fetch('document/update', {
         body: JSON.stringify(Object.assign({}, dataItem, item)),
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
       });
       message.success(`${item.category} 更新完成`);
     } else {
-      const response = await fetch('save/new', {
+      const response = await fetch('document/add', {
         body: JSON.stringify(item),
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
